--- internal/remote_coder/bot/manager_test.go
@@ -1,137 +1,0 @@
-package bot
-
-import (
-	"context"
-	"path/filepath"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/tingly-dev/tingly-box/agentboot"
-	"github.com/tingly-dev/tingly-box/internal/data/db"
-	"github.com/tingly-dev/tingly-box/internal/remote_coder/session"
-)
-
-func TestManagerStartStop(t *testing.T) {
-	dir := t.TempDir()
-	dbPath := filepath.Join(dir, "tingly.db")
-
-	store, err := NewStore(dbPath)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = store.Close() })
-
-	// Create a session manager
-	msgStore, err := session.NewMessageStore(dbPath)
-	require.NoError(t, err)
-	sessionMgr := session.NewManager(session.Config{
-		Timeout:          30 * time.Minute,
-		MessageRetention: 24 * time.Hour,
-	}, msgStore)
-
-	// Create agentBoot and permission handler for test
-	agentBoot := agentboot.New(agentboot.Config{})
-
-	// Create manager
-	botStore, err := db.NewImBotSettingsStore(dir)
-	manager := NewManager(botStore, sessionMgr, agentBoot)
-	manager.SetDataPath(filepath.Join(dir, "bot_chats.json"))
-
-	// Test: IsRunning returns false for non-existent bot
-	require.False(t, manager.IsRunning("non-existent-uuid"))
-
-	// Test: Stop on non-running bot is safe (no panic)
-	manager.Stop("non-existent-uuid")
-}
-
-func TestManagerStartEnabledBots(t *testing.T) {
-	dir := t.TempDir()
-	dbPath := filepath.Join(dir, "tingly.db")
-
-	store, err := NewStore(dbPath)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = store.Close() })
-
-	// Create a session manager
-	msgStore, err := session.NewMessageStore(dbPath)
-	require.NoError(t, err)
-	sessionMgr := session.NewManager(session.Config{
-		Timeout:          30 * time.Minute,
-		MessageRetention: 24 * time.Hour,
-	}, msgStore)
-
-	botStore, err := db.NewImBotSettingsStore(dir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create enabled bot (without token - won't actually run)
-	_, err = botStore.CreateSettings(db.Settings{
-		Name:     "Test Bot",
-		Platform: "telegram",
-		AuthType: "token",
-		Auth:     map[string]string{}, // No token
-		Enabled:  true,
-	})
-	require.NoError(t, err)
-
-	// Create disabled bot
-	_, err = botStore.CreateSettings(db.Settings{
-		Name:     "Disabled Bot",
-		Platform: "telegram",
-		AuthType: "token",
-		Auth:     map[string]string{"token": "test-token"},
-		Enabled:  false,
-	})
-	require.NoError(t, err)
-
-	// Create agentBoot and permission handler for test
-	agentBoot := agentboot.New(agentboot.Config{})
-
-	// Create manager
-	manager := NewManager(botStore, sessionMgr, agentBoot)
-	manager.SetDataPath(filepath.Join(dir, "bot_chats.json"))
-
-	// Start enabled bots - should not fail
-	ctx := context.Background()
-	err = manager.StartEnabled(ctx)
-	require.NoError(t, err)
-
-	// Give time for goroutines to process
-	time.Sleep(100 * time.Millisecond)
-
-	// Bot without token should not be running
-	settings, err := botStore.ListEnabledSettings()
-	require.NoError(t, err)
-	if len(settings) > 0 {
-		// The bot without token won't start
-		require.False(t, manager.IsRunning(settings[0].UUID))
-	}
-}
-
-func TestManagerStopAll(t *testing.T) {
-	dir := t.TempDir()
-	dbPath := filepath.Join(dir, "tingly.db")
-
-	store, err := NewStore(dbPath)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = store.Close() })
-
-	msgStore, err := session.NewMessageStore(dbPath)
-	require.NoError(t, err)
-	sessionMgr := session.NewManager(session.Config{
-		Timeout:          30 * time.Minute,
-		MessageRetention: 24 * time.Hour,
-	}, msgStore)
-
-	// Create agentBoot and permission handler for test
-	agentBoot := agentboot.New(agentboot.Config{})
-
-	// Create manager
-	botStore, err := db.NewImBotSettingsStore(dir)
-	manager := NewManager(botStore, sessionMgr, agentBoot)
-	manager.SetDataPath(filepath.Join(dir, "bot_chats.json"))
-
-	// StopAll should be safe even with no running bots
-	manager.StopAll()
-}
