--- internal/remote_coder/bot/store_test.go
@@ -1,66 +1,0 @@
-package bot
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"testing"
-
-	"github.com/google/uuid"
-	"github.com/stretchr/testify/require"
-	"github.com/tingly-dev/tingly-box/internal/data/db"
-)
-
-func TestStoreSettingsRoundTrip(t *testing.T) {
-	dir := t.TempDir()
-	dbPath := filepath.Join(dir, "tingly.db")
-	store, err := NewStore(dbPath)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = store.Close() })
-
-	uid, err := uuid.NewUUID()
-	settings := db.Settings{
-		UUID:          uid.String(),
-		Token:         "telegram-token",
-		Platform:      "telegram",
-		ProxyURL:      "http://proxy.test:8080",
-		ChatIDLock:    "chat-123",
-		BashAllowlist: []string{"cd", "ls", "pwd"},
-		Auth: map[string]string{
-			"token": "telegram-token",
-		},
-	}
-
-	botStore, err := db.NewImBotSettingsStore(dir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	_, err = botStore.CreateSettings(settings)
-	require.NoError(t, err)
-
-	loaded, err := botStore.GetSettingsByUUID(uid.String())
-	fmt.Printf("loaded: %v", loaded)
-	require.NoError(t, err)
-	require.Equal(t, "telegram-token", loaded.Token)
-	require.Equal(t, "telegram", loaded.Platform)
-	require.Equal(t, "http://proxy.test:8080", loaded.ProxyURL)
-	require.Equal(t, "chat-123", loaded.ChatIDLock)
-	require.Equal(t, []string{"cd", "ls", "pwd"}, loaded.BashAllowlist)
-
-	_, err = os.Stat(dbPath)
-	require.NoError(t, err)
-}
-
-func TestStoreSessionMapping(t *testing.T) {
-	dir := t.TempDir()
-	store, err := NewStore(filepath.Join(dir, "tingly.db"))
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = store.Close() })
-
-	require.NoError(t, store.SetSessionForChat("chat-1", "session-1"))
-	id, ok, err := store.GetSessionForChat("chat-1")
-	require.NoError(t, err)
-	require.True(t, ok)
-	require.Equal(t, "session-1", id)
-}
